Fix `clippy::drop_ref` lint
Clippy doesn't like calls to `std::mem::drop` with references --- since
the pointed value is not actually dropped. In this case, this *is* the
correct behavior, but clippy views this as a footgun if the user means
to actually drop the value behind the reference.

Here, we're using `drop` to ignore values in no-op default impls
for trait methods.  I've
changed those methods to use `let _` to ignore parameters instead. This
doesn't trigger the clippy lint and is maybe more idiomatic anyway.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>