silence warning on non-epsilon float comparisons
The tests will make assertions that particular values are recorded in a
particular order. When testing with floating-point values, clippy emits
a warning that we should be using epsilon comparisons rather than
bit equality. However, this isn't an issue here, since the tests just using
float literals, and if the right float is recorded in the right order,
they will be precisely equal.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>