Add `Slice::is_empty()` method
Clippy lints about types which have a public `len` method but no
corresponding `is_empty` method. Therefore, I've added an `is_empty`
method that returns `true` if `self.len() > 0`.

Since this is a new API that we'll have to support, it would also be
fine to solve this by _not_ adding the method and just suppressing the
warning...but it seems worth having for consistency with the underlying
slice's methods? And, I don't see how this could really pose a
forward-compatibility hazard, since `Slice` is...always going to be
backed by an actual slice...