fields: Add the number of fields to `Fields::Unnamed`
Closes #36

Currently, there is no way to determine the number of fields in a tuple
struct or tuple-like enum variant without actually visiting the fields.
On the other hand, it _is_ possible to get the number of fields from a
struct or struct-like enum variant, by calling `.len()` on the slice of
`NamedFields` in `Fields::Named`.

It seems like both `Fields::Unnamed` _and_ `Fields::Named` probably
ought to have a way to determine the number of fields. I can imagine
use-cases where (say) a serializer wants to know the number of fields
before actually visiting them, in order to pre-allocate storage, or emit
a length delimiter, or something.

This branch changes `Fields::Unnamed` to store a `usize` number of
fields. It also adds `Fields::len` and `Fields::is_empty` methods to
`Fields`, allowing the length to be accessed without having to
destructure the enum.

Since this adds a field to a public enum, it's a breaking change --- we
probably want to get this in for 1.0.