add `doc(cfg(...))` to feature-flagged impls
Currently, feature-flagged public APIs will not be documented as
requiring feature flags. This branch adds `#[doc(cfg(...))]` to all
feature-flagged APIs.

For the most part, this is done using the `feature!` macro lifted from
Tokio:
https://github.com/tokio-rs/tokio/blob/8943e8aeef0b33f371d6dc69f62b38da390b5d5f/tokio/src/macros/cfg.rs#L3-L14

In a couple of places, though, the feature-flagged thing is in a
position other than item position, so in that case, it was necessary to
just paste the attribute...

Signed-off-by: Eliza Weisman <eliza@buoyant.io>