valuable-serde: prepare to release v0.1.0 (#83)
This branch prepares the `valuable-serde` crate for an initial crates.io
release. I've expanded on the docs a little, and added a summary of
`valuable` and a link to its own documentation, so that anyone who reads
the docs for `valuable-serde` can find the rest of the `valuable`
ecosystem. I also added a README, an initial changelog, and some missing
`Cargo.toml` metadata.

I've also added a "Related Crates" section in `valuable`'s `lib.rs` docs
that links to `valuable-serde`, so that users can know it exists.

Closes #82

Signed-off-by: Eliza Weisman <eliza@buoyant.io>